pyproject.toml
[project]
name = "hospital-client"
version = "0.1.0"
description = ""
authors = [
    {name = "Mayuresh Anand",email = "56501614+mayureshucsb2019@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "pyaudio (>=0.2.14,<0.3.0)",
    "sounddevice (>=0.5.1,<0.6.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "assemblyai (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "pdfkit (>=1.0.0,<2.0.0)",
    "markdown (>=3.7,<4.0)",
    "wkhtmltopdf (>=0.2,<0.3)",
    "weasyprint (>=65.0,<66.0)",
    "pyannote-audio (>=3.3.2,<4.0.0)",
    "transformers (>=4.51.1,<5.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "autoflake (>=2.3.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "hospital_client", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
